requirements.txt
# Core Dependencies
groq==0.9.0
python-dotenv==1.0.0

# RAG and Vector Database
chromadb==0.4.22
langchain==0.1.20
langchain-community==0.0.38
sentence-transformers==2.5.1

# Web Search and Document Processing
requests==2.31.0
beautifulsoup4==4.12.3
pypdf==4.0.1
python-docx==1.1.0

# Data Processing
numpy==1.26.4
pandas==2.2.0

# Additional Tools
tiktoken==0.6.0
pydantic==2.6.1
colorama==0.4.6
